pyproject.toml
[project]
name = "fastapi-session"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=5.29.0",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "pathspec>=0.12.1",
    "pyjwt>=2.10.1",
]

[tool.ruff]
lint.select = ["I"]   # "I" is the code for isort rules (import sorting)

[tool.ruff.lint.isort]
known-first-party = ["fastapi-session"]  # Optional: your local module names
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
force-sort-within-sections = true     # Optional: sort within groups
combine-as-imports = true             # Optional: combine 'import a as b'
